project/src/edge-compute/requirements.txt
openai-whisper
scikit-learn
numpy
setuptools-rust